+++ java/net/sf/eventengine/events/AllVsAll.java
@@ -20,23 +20,19 @@ package net.sf.eventengine.events;
 
 import java.util.List;
 
-import com.l2jserver.gameserver.model.actor.L2Character;
-import com.l2jserver.gameserver.model.actor.L2Npc;
-import com.l2jserver.gameserver.model.instancezone.InstanceWorld;
-import com.l2jserver.gameserver.model.items.L2Item;
-import com.l2jserver.gameserver.model.skills.Skill;
-import com.l2jserver.gameserver.network.clientpackets.Say2;
-
 import net.sf.eventengine.datatables.ConfigData;
 import net.sf.eventengine.datatables.MessageData;
 import net.sf.eventengine.enums.CollectionTarget;
-import net.sf.eventengine.enums.EventState;
 import net.sf.eventengine.enums.TeamType;
 import net.sf.eventengine.events.handler.AbstractEvent;
 import net.sf.eventengine.events.holders.PlayerHolder;
 import net.sf.eventengine.events.schedules.AnnounceNearEndEvent;
 import net.sf.eventengine.util.EventUtil;
 import net.sf.eventengine.util.SortUtil;
+
+import com.l2jserver.gameserver.model.actor.L2Character;
+import com.l2jserver.gameserver.model.instancezone.InstanceWorld;
+import com.l2jserver.gameserver.network.clientpackets.Say2;
 
 /**
  * @author fissban
@@ -59,26 +55,22 @@ public class AllVsAll extends AbstractEvent
 	}
 	
 	@Override
-	public void runEventState(EventState state)
+	protected void onEventStart()
 	{
-		switch (state)
-		{
-			case START:
-				prepareToStart(); // General Method
-				createTeam(1);
-				teleportAllPlayers(RADIUS_SPAWN_PLAYER);
-				break;
-				
-			case FIGHT:
-				prepareToFight(); // General Method
-				break;
-				
-			case END:
-				giveRewardsTeams();
-				prepareToEnd(); // General Method
-				break;
-		}
-		
+		createTeam(1);
+		teleportAllPlayers(RADIUS_SPAWN_PLAYER);
+	}
+	
+	@Override
+	protected void onEventFight()
+	{
+		// Nothing
+	}
+	
+	@Override
+	protected void onEventEnd()
+	{
+		giveRewardsTeams();
 	}
 	
 	// LISTENERS -----------------------------------------------------------------------
@@ -117,18 +109,6 @@ public class AllVsAll extends AbstractEvent
 	}
 	
 	@Override
-	public boolean onAttack(PlayerHolder ph, L2Character target)
-	{
-		return false;
-	}
-	
-	@Override
-	public boolean onUseSkill(PlayerHolder ph, L2Character target, Skill skill)
-	{
-		return false;
-	}
-	
-	@Override
 	public void onDeath(PlayerHolder ph)
 	{
 		// We generated a task to revive the player
@@ -137,24 +117,6 @@ public class AllVsAll extends AbstractEvent
 		ph.increaseDeaths();
 		// We update the title character
 		updateTitle(ph);
-	}
-	
-	@Override
-	public boolean onInteract(PlayerHolder ph, L2Npc npc)
-	{
-		return true;
-	}
-	
-	@Override
-	public boolean onUseItem(PlayerHolder ph, L2Item item)
-	{
-		return false;
-	}
-	
-	@Override
-	public void onLogout(PlayerHolder ph)
-	{
-		//
 	}
 	
 	// VARIOUS METHODS ------------------------------------------------------------------
