+++ java/net/sf/eventengine/events/OneVsOne.java
@@ -24,24 +24,20 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import com.l2jserver.gameserver.ThreadPoolManager;
-import com.l2jserver.gameserver.model.actor.L2Character;
-import com.l2jserver.gameserver.model.actor.L2Npc;
-import com.l2jserver.gameserver.model.instancezone.InstanceWorld;
-import com.l2jserver.gameserver.model.items.L2Item;
-import com.l2jserver.gameserver.model.skills.Skill;
-
 import net.sf.eventengine.EventEngineManager;
 import net.sf.eventengine.datatables.ConfigData;
 import net.sf.eventengine.datatables.MessageData;
 import net.sf.eventengine.enums.EventEngineState;
-import net.sf.eventengine.enums.EventState;
 import net.sf.eventengine.enums.TeamType;
 import net.sf.eventengine.events.handler.AbstractEvent;
 import net.sf.eventengine.events.holders.PlayerHolder;
 import net.sf.eventengine.events.schedules.AnnounceNearEndEvent;
 import net.sf.eventengine.util.EventUtil;
 import net.sf.eventengine.util.SortUtil;
+
+import com.l2jserver.gameserver.ThreadPoolManager;
+import com.l2jserver.gameserver.model.actor.L2Character;
+import com.l2jserver.gameserver.model.instancezone.InstanceWorld;
 
 /**
  * @author fissban
@@ -65,40 +61,25 @@ public class OneVsOne extends AbstractEvent
 	}
 	
 	@Override
-	public void runEventState(EventState state)
+	protected void onEventStart()
 	{
-		switch (state)
-		{
-			case START:
-				prepareToStart(); // General Method
-				createTeams(ConfigData.getInstance().OVO_COUNT_TEAM);
-				teleportAllPlayers(RADIUS_SPAWN_PLAYER);
-				break;
-				
-			case FIGHT:
-				prepareToFight(); // General Method
-				break;
-				
-			case END:
-				giveRewardsTeams();
-				prepareToEnd(); // General Method
-				break;
-		}
+		createTeams(ConfigData.getInstance().OVO_COUNT_TEAM);
+		teleportAllPlayers(RADIUS_SPAWN_PLAYER);
+	}
+	
+	@Override
+	protected void onEventFight()
+	{
+		// Nothing
+	}
+	
+	@Override
+	protected void onEventEnd()
+	{
+		giveRewardsTeams();
 	}
 	
 	// LISTENERS ------------------------------------------------------
-	@Override
-	public boolean onUseSkill(PlayerHolder ph, L2Character target, Skill skill)
-	{
-		return false;
-	}
-	
-	@Override
-	public boolean onAttack(PlayerHolder ph, L2Character target)
-	{
-		return false;
-	}
-	
 	@Override
 	public void onKill(PlayerHolder ph, L2Character target)
 	{
@@ -148,30 +129,6 @@ public class OneVsOne extends AbstractEvent
 		{
 			EventUtil.messageKill(ph, target);
 		}
-	}
-	
-	@Override
-	public void onDeath(PlayerHolder ph)
-	{
-		//
-	}
-	
-	@Override
-	public boolean onInteract(PlayerHolder ph, L2Npc npc)
-	{
-		return true;
-	}
-	
-	@Override
-	public boolean onUseItem(PlayerHolder ph, L2Item item)
-	{
-		return false;
-	}
-	
-	@Override
-	public void onLogout(PlayerHolder ph)
-	{
-		//
 	}
 	
 	// VARIOUS METHODS -------------------------------------------------
@@ -285,7 +242,7 @@ public class OneVsOne extends AbstractEvent
 				}
 			}
 			
-		} , TIME_BETWEEN_FIGHT * 1000);
+		}, TIME_BETWEEN_FIGHT * 1000);
 	}
 	
 	/**
